設定画面からtokenを登録したときもエラーだった場合はアラートを表示する
